requirements.txt
llama-cloud==0.1.4
llama-index==0.11.19
llama-index-agent-openai==0.3.4
llama-index-cli==0.3.1
llama-index-core==0.11.19
llama-index-embeddings-openai==0.2.5
llama-index-indices-managed-llama-cloud==0.4.0
llama-index-legacy==0.9.48.post3
llama-index-llms-openai==0.2.16
llama-index-multi-modal-llms-openai==0.2.3
llama-index-program-openai==0.2.0
llama-index-question-gen-openai==0.2.0
llama-index-readers-file==0.2.2
llama-index-readers-llama-parse==0.3.0
llama-parse==0.5.11
chromadb==0.5.15
llama-index-llms-gemini==0.3.7
llama-index-embeddings-gemini==0.2.2
google-generativeai==0.5.4
llama-index-vector-stores-chroma==0.2.2
python-dotenv==1.0.1
transformers==4.46.1
torch==2.2.1+cu121
pydub==0.25.1
streamlit